debug the problem of prediction

+++ mcls-main/samcl/SAMCL.java
@@ -503,7 +503,7 @@ public class SAMCL implements Closeable{
 						if(i>10)
 							continue first;
 						i++;
-						Distribution.MotionSampling(p, u, duration/1000, new Random(), this.al); 
+						Distribution.MotionSampling(p, u, (double)duration/1000, new Random(), this.al); 
 					}while(!Distribution.boundaryCheck(p, this.grid));
 					
 					result.add(p.clone());
